chore(web): inline single-use mutation wrappers
Mutation hooks consumed at a single call site are inlined per the
documented preference. Wrappers that hold the only export are
deleted; useArchiveWorkspace had zero consumers and is removed.
useAddWorkspaceMember is kept (two call sites).
