--- apps/web/src/routes/_protected.workspaces/$workspaceId/-mutations/billing-codes.ts
@@ -1,103 +1,0 @@
-import { useMutation } from "@tanstack/react-query";
-
-import { useAnalytics } from "@/lib/analytics/provider";
-import { api } from "@/lib/api";
-import { toAPIError } from "@/lib/errors";
-import { toSafeId } from "@/lib/safe-id";
-import { billingCodesKeys } from "@/routes/_protected.workspaces/$workspaceId/-queries/billing-codes";
-
-type CreateBillingCodeVars = {
-  workspaceId: string;
-  type: "task" | "activity";
-  code: string;
-  label: string;
-  active?: boolean;
-  sortOrder?: number;
-};
-
-export const useCreateBillingCode = () => {
-  const analytics = useAnalytics();
-
-  return useMutation({
-    mutationFn: async ({ workspaceId, ...body }: CreateBillingCodeVars) => {
-      const response = await api["billing-codes"]({
-        workspaceId: toSafeId<"workspace">(workspaceId),
-      }).put({
-        queryKey: billingCodesKeys.all(workspaceId),
-        ...body,
-      });
-
-      if (response.error) {
-        throw toAPIError(response.error);
-      }
-
-      return response.data;
-    },
-    onError: (error) => {
-      analytics.captureError(error);
-    },
-  });
-};
-
-type UpdateBillingCodeVars = {
-  workspaceId: string;
-  id: string;
-  code?: string;
-  label?: string;
-  active?: boolean;
-  sortOrder?: number;
-};
-
-export const useUpdateBillingCode = () => {
-  const analytics = useAnalytics();
-
-  return useMutation({
-    mutationFn: async ({ workspaceId, ...body }: UpdateBillingCodeVars) => {
-      const response = await api["billing-codes"]({
-        workspaceId: toSafeId<"workspace">(workspaceId),
-      }).patch({
-        queryKey: billingCodesKeys.all(workspaceId),
-        ...body,
-        id: toSafeId<"billingCode">(body.id),
-      });
-
-      if (response.error) {
-        throw toAPIError(response.error);
-      }
-
-      return response.data;
-    },
-    onError: (error) => {
-      analytics.captureError(error);
-    },
-  });
-};
-
-type DeleteBillingCodeVars = {
-  workspaceId: string;
-  id: string;
-};
-
-export const useDeleteBillingCode = () => {
-  const analytics = useAnalytics();
-
-  return useMutation({
-    mutationFn: async ({ workspaceId, id }: DeleteBillingCodeVars) => {
-      const response = await api["billing-codes"]({
-        workspaceId: toSafeId<"workspace">(workspaceId),
-      }).delete({
-        queryKey: billingCodesKeys.all(workspaceId),
-        id: toSafeId<"billingCode">(id),
-      });
-
-      if (response.error) {
-        throw toAPIError(response.error);
-      }
-
-      return response.data;
-    },
-    onError: (error) => {
-      analytics.captureError(error);
-    },
-  });
-};
